# rest/model.go
package rest

import "encoding/json"

type SortType int

const (
	sortTypeAsc SortType = iota
	sortTypeDesc
	sortTypeNone
)

type SortFieldType int

const (
	sortFieldTypeID SortFieldType = iota
	sortFieldTypeHExId
	sortFieldTypeFrom
	sortFieldTypeTo
	sortFieldTypeWeight
)

type GetListRequest struct {
	Filter map[string]any `form:"filter"`
	Range  string         `form:"range"`
	Sort   string         `form:"sort"`
}

type CreateNodeRequest struct {
	ID    uint   `json:"id"`
	Tag   string `json:"tag"`
	InUse bool   `json:"in_use"`
}

type UpdateNodeRequest struct {
	Tag     string `json:"tag"`
	InUse   bool   `json:"in_use"`
	DevTag  string `json:"dev_tag"`
	Channel int8   `json:"channel"`
	TxPower int8   `json:"tx_power"`
}

type MeshNode struct {
	ID       uint   `json:"id"`
	Tag      string `json:"tag"`
	InUse    bool   `json:"in_use"`
	Path     string `json:"path"`
	Revision string `json:"revision"`
	Error    string `json:"error"`
	DevTag   string `json:"dev_tag"`
	Channel  int8   `json:"channel"`
	TxPower  int8   `json:"tx_power"`
	Groups   int    `json:"groups"`
	Binded   int    `json:"binded"`
	Flags    int    `json:"flags"`
}

type UpdateLinkRequest struct {
	ID     uint    `json:"id"`
	Weight float32 `json:"weight"`
}

type MeshLink struct {
	ID     uint    `json:"id"`
	From   uint    `json:"from"`
	To     uint    `json:"to"`
	Weight float32 `json:"weight"`
}

type MeshNeighbor struct {
	ID      uint    `json:"id"`
	Address string  `json:"address"`
	Current float32 `json:"current"`
	Next    float32 `json:"next"`
	Delta   float32 `json:"delta"`
}

type MeshDiscoveryState struct {
	ID        int64  `json:"id"`
	Status    string `json:"status"`
	CurrentId int64  `json:"current_id"`
	Repeat    int    `json:"repeat"`
}

type MeshFirmware struct {
	ID       int64  `json:"id"`
	Status   string `json:"status"`
	Filename string `json:"filename"`
	Size     int64  `json:"size"`
	Progress int    `json:"progress"`
}

type UpdateFirmwareRequest struct {
	ID       int64  `json:"id"`
	Filename string `json:"filename"`
}

var acceptFilters = map[string]struct{}{
	"from": {},
	"to":   {},
	"any":  {},
}

type EsphomeServer struct {
	ID      uint   `json:"id"`
	Address string `json:"address"`
	Clients int    `json:"clients"`
}

type EsphomeClient struct {
	ID       uint   `json:"id"`
	Address  string `json:"address"`
	Tag      string `json:"tag"`
	Active   bool   `json:"active"`
	Handle   int    `json:"handle"`
	Sent     int    `json:"sent"`
	Received int    `json:"received"`
	Duration string `json:"duration"`
	Started  string `json:"started"`
}

type GetListParams struct {
	Filter        map[string]interface{}
	Limit, Offset int
	SortBy        SortFieldType
	SortType      SortType
}

func parseSortType(s string) SortType {
	switch s {
	case "ASC":
		return sortTypeAsc
	case "DESC":
		return sortTypeDesc
	}
	return sortTypeAsc
}

func parseSortFieldType(s string) SortFieldType {
	switch s {
	case "id":
		return sortFieldTypeID
	case "hex_id":
		return sortFieldTypeHExId
	case "from":
		return sortFieldTypeFrom
	case "to":
		return sortFieldTypeTo
	case "weight":
		return sortFieldTypeWeight
	}
	return sortFieldTypeID
}

func (r GetListRequest) toGetListParams() GetListParams {
	acceptFiltersParam := make(map[string]any)
	for k, v := range r.Filter {
		if _, ok := acceptFilters[k]; ok {
			acceptFiltersParam[k] = v
		}
	}

	p := GetListParams{
		Filter:   acceptFiltersParam,
		Limit:    10,
		Offset:   0,
		SortBy:   sortFieldTypeID,
		SortType: sortTypeAsc,
	}

	if r.Range != "" {
		var queryRange []int
		_ = json.Unmarshal([]byte(r.Range), &queryRange)
		if len(queryRange) == 2 {
			p.Offset, p.Limit = queryRange[0], queryRange[1]
		}
	}

	if r.Sort != "" {
		var querySort []string
		_ = json.Unmarshal([]byte(r.Sort), &querySort)
		if len(querySort) == 2 {
			p.SortBy, p.SortType = parseSortFieldType(querySort[0]), parseSortType(querySort[1])
		}
	}

	return p
}